+++ src/parsing/lexer/ret_value.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ret_value.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: efrancon <efrancon@student.42.fr>          +#+  +:+       +#+        */
+/*   By: lraffin <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/04 14:47:11 by efrancon          #+#    #+#             */
-/*   Updated: 2021/11/06 15:02:15 by efrancon         ###   ########.fr       */
+/*   Updated: 2021/11/11 17:54:16 by lraffin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,7 +49,7 @@ static int	get_length(char *str, int value_length)
 	i = 0;
 	count = 0;
 	length = ft_strlen(str);
-	while (i < (int)ft_strlen(str) && str[i] && str[i + 1])
+	while (i < ft_strlen(str) && str[i] && str[i + 1])
 	{
 		if (str[i] && (str[i] == '\'' || str[i] == '\"'))
 			get_length_quotes(&length, &count, &i, str);
